Use atol

+++ examples/all-clusters-app/linux/AppOptions.cpp
@@ -148,7 +148,7 @@ bool AppOptions::HandleOptions(const char * program, OptionSet * options, int id
             using namespace chip::app::Clusters::TimeSynchronization;
             ForceTimeSource(TimeSourceEnum::kUnknown);
         }
-        long longValue = atoi(value);
+        long longValue = atol(value);
         if (longValue >= 0)
         {
             uint64_t override = uint64_t(longValue) * chip::kMicrosecondsPerSecond;
